+++ lib/stringr.dart
@@ -1,6 +1,5 @@
 library stringr;
 
-export 'src/case/case.dart';
 export 'src/util/strings/strings.dart';
 export 'src/util/strings/surrogate_pair.dart';
 export 'src/util/object/extended_iterable.dart';
@@ -10,3 +9,4 @@ export 'src/split/words.dart';
 export 'src/count/count.dart';
 export 'src/escape/escape.dart';
 export 'src/manipulate/manipulate.dart';
+export 'src/case/case.dart';